refactor: moved TncToGtfsWriter to io.
